[fix]ScrollView and ListView goes back to the front incorrectly after calling doLayout.

+++ cocos/2d/CCFontAtlas.cpp
@@ -376,8 +376,8 @@ bool FontAtlas::prepareLetterDefinitions(const std::u32string& utf32Text)
 
     int adjustForDistanceMap = _letterPadding / 2;
     int adjustForExtend = _letterEdgeExtend / 2;
-    int32_t bitmapWidth;
-    int32_t bitmapHeight;
+    long bitmapWidth;
+    long bitmapHeight;
     int glyphHeight;
     Rect tempRect;
     FontLetterDefinition tempDef;
