+++ cocos/renderer/CCTextureCache.cpp
@@ -652,14 +652,14 @@ std::string TextureCache::getCachedTextureInfo() const
         auto bytes = tex->getPixelsWide() * tex->getPixelsHigh() * bpp / 8;
         totalBytes += bytes;
         count++;
-        snprintf(buftmp, sizeof(buftmp) - 1, "\"%s\" rc=%d id=%p %d x %d @ %d bpp => %d KB\n",
+        snprintf(buftmp, sizeof(buftmp) - 1, "\"%s\" rc=%lu id=%p %lu x %lu @ %ld bpp => %lu KB\n",
             texture.first.c_str(),
-            (int32_t)tex->getReferenceCount(),
+            (long)tex->getReferenceCount(),
             tex->getBackendTexture(),
-            (int32_t)tex->getPixelsWide(),
-            (int32_t)tex->getPixelsHigh(),
-            (int32_t)bpp,
-            (int32_t)bytes / 1024);
+            (long)tex->getPixelsWide(),
+            (long)tex->getPixelsHigh(),
+            (long)bpp,
+            (long)bytes / 1024);
 
         buffer += buftmp;
     }
